// demo/src/lib/fund-engine/generated/model-inputs.ts
export const GENERATED_MODEL_INPUT_FIELDS = [
  {
    "id": "fundTerm",
    "label": "Life of Fund",
    "group": "fund",
    "row": 4,
    "column": 5,
    "valueType": "number",
    "defaultValue": 15,
    "unit": "count",
    "minimum": 0,
    "integer": true,
    "maximum": 15
  },
// ...
] as const